+++ lib/resource/template/date/range/DateRangePickerResult.dart
@@ -1,0 +1,12 @@
+class DateRangePickerResult {
+
+  String start;
+  String end;
+
+  DateRangePickerResult(this.start, this.end);
+
+  @override
+  String toString() {
+    return 'DateRangePickerResult{start: $start, end: $end}';
+  }
+}